migrating video to frames conversion to core module
